+++ src/events/events.service.spec.ts
@@ -220,4 +220,5 @@ describe('EventsService', () => {
       NotFoundException,
     );
   });
+
 });