+++ src/media/types/admin-query-params.interface.ts
@@ -1,10 +1,11 @@
 export interface MediaAdminQueryParams {
   limit?: string;
   offset?: string;
+  isPublished?: string | boolean;
   kind?: string;
   search?: string;
   folder?: string;
   createdFrom?: string;
   createdTo?: string;
-  [key: string]: string | string[] | undefined;
+  [key: string]: string | string[] | boolean | undefined;
 }