+++ src/people/people.service.spec.ts
@@ -301,4 +301,5 @@ describe('PeopleService', () => {
       NotFoundException,
     );
   });
+
 });